+++ dio/lib/src/dio_mixin.dart
@@ -341,7 +341,7 @@ abstract class DioMixin implements Dio {
     Options? options,
     ProgressCallback? onSendProgress,
     ProgressCallback? onReceiveProgress,
-  }) async {
+  }) {
     final requestOptions = (options ?? Options()).compose(
       this.options,
       path,
@@ -378,61 +378,59 @@ abstract class DioMixin implements Dio {
     // Convert the request interceptor to a functional callback in which
     // we can handle the return value of interceptor callback.
     FutureOr Function(dynamic) requestInterceptorWrapper(
-      InterceptorSendCallback interceptor,
+      InterceptorSendCallback cb,
     ) {
-      return (dynamic incomingState) async {
+      return (dynamic incomingState) {
         final state = incomingState as InterceptorState;
         if (state.type == InterceptorResultType.next) {
           return listenCancelForAsyncTask(
             requestOptions.cancelToken,
-            Future(() {
-              final requestHandler = RequestInterceptorHandler();
-              interceptor(state.data as RequestOptions, requestHandler);
-              return requestHandler.future;
+            Future(() async {
+              final handler = RequestInterceptorHandler();
+              cb(state.data as RequestOptions, handler);
+              return handler.future;
             }),
           );
-        } else {
-          return state;
         }
+        return state;
       };
     }
 
     // Convert the response interceptor to a functional callback in which
     // we can handle the return value of interceptor callback.
     FutureOr<dynamic> Function(dynamic) responseInterceptorWrapper(
-      InterceptorSuccessCallback interceptor,
+      InterceptorSuccessCallback cb,
     ) {
-      return (dynamic incomingState) async {
+      return (dynamic incomingState) {
         final state = incomingState as InterceptorState;
         if (state.type == InterceptorResultType.next ||
             state.type == InterceptorResultType.resolveCallFollowing) {
           return listenCancelForAsyncTask(
             requestOptions.cancelToken,
-            Future(() {
-              final responseHandler = ResponseInterceptorHandler();
-              interceptor(state.data as Response, responseHandler);
-              return responseHandler.future;
+            Future(() async {
+              final handler = ResponseInterceptorHandler();
+              cb(state.data as Response, handler);
+              return handler.future;
             }),
           );
-        } else {
-          return state;
         }
+        return state;
       };
     }
 
     // Convert the error interceptor to a functional callback in which
     // we can handle the return value of interceptor callback.
     FutureOr<dynamic> Function(Object) errorInterceptorWrapper(
-      InterceptorErrorCallback interceptor,
+      InterceptorErrorCallback cb,
     ) {
-      return (error) {
+      return (dynamic error) {
         final state = error is InterceptorState
             ? error
             : InterceptorState(assureDioException(error, requestOptions));
         Future<InterceptorState> handleError() async {
-          final errorHandler = ErrorInterceptorHandler();
-          interceptor(state.data, errorHandler);
-          return errorHandler.future;
+          final handler = ErrorInterceptorHandler();
+          cb(state.data, handler);
+          return handler.future;
         }
 
         // The request has already been cancelled,
@@ -440,15 +438,15 @@ abstract class DioMixin implements Dio {
         if (state.data is DioException &&
             state.data.type == DioExceptionType.cancel) {
           return handleError();
-        } else if (state.type == InterceptorResultType.next ||
+        }
+        if (state.type == InterceptorResultType.next ||
             state.type == InterceptorResultType.rejectCallFollowing) {
           return listenCancelForAsyncTask(
             requestOptions.cancelToken,
             Future(handleError),
           );
-        } else {
-          throw error;
         }
+        throw error;
       };
     }
 
@@ -471,13 +469,14 @@ abstract class DioMixin implements Dio {
       requestInterceptorWrapper((
         RequestOptions reqOpt,
         RequestInterceptorHandler handler,
-      ) {
+      ) async {
         requestOptions = reqOpt;
-        _dispatchRequest<T>(reqOpt)
-            .then((value) => handler.resolve(value, true))
-            .catchError((e) {
-          handler.reject(e as DioException, true);
-        });
+        try {
+          final value = await _dispatchRequest<T>(reqOpt);
+          handler.resolve(value, true);
+        } on DioException catch (e) {
+          handler.reject(e, true);
+        }
       }),
     );
 
@@ -497,12 +496,13 @@ abstract class DioMixin implements Dio {
       future = future.catchError(errorInterceptorWrapper(fun));
     }
     // Normalize errors, converts errors to [DioException].
-    return future.then<Response<T>>((data) {
+    try {
+      final data = await future;
       return assureResponse<T>(
         data is InterceptorState ? data.data : data,
         requestOptions,
       );
-    }).catchError((Object e) {
+    } catch (e) {
       final isState = e is InterceptorState;
       if (isState) {
         if (e.type == InterceptorResultType.resolve) {
@@ -510,7 +510,7 @@ abstract class DioMixin implements Dio {
         }
       }
       throw assureDioException(isState ? e.data : e, requestOptions);
-    });
+    }
   }
 
   Future<Response<dynamic>> _dispatchRequest<T>(RequestOptions reqOpt) async {
@@ -682,10 +682,10 @@ abstract class DioMixin implements Dio {
     CancelToken? cancelToken,
     Future<T> future,
   ) {
-    return Future.any([
-      if (cancelToken != null) cancelToken.whenCancel.then((e) => throw e),
-      future,
-    ]);
+    if (cancelToken == null) {
+      return future;
+    }
+    return Future.any([future, cancelToken.whenCancel.then((e) => throw e)]);
   }
 
   @internal