+++ dio_test/lib/src/test/suite.dart
@@ -1,0 +1,27 @@
+import 'package:dio/dio.dart';
+import 'package:dio_test/tests.dart';
+
+typedef TestSuiteFunction = void Function(
+  Dio Function(String baseUrl) create,
+);
+
+const _tests = [
+  basicTests,
+  cancellationTests,
+  corsTests,
+  downloadTests,
+  headerTests,
+  httpMethodTests,
+  parameterTests,
+  redirectTests,
+  statusCodeTests,
+  timeoutTests,
+  uploadTests,
+  urlEncodedTests,
+];
+
+void dioAdapterTestSuite(
+  Dio Function(String baseUrl) create, {
+  List<TestSuiteFunction> tests = _tests,
+}) =>
+    tests.forEach((test) => test(create));