+++ plugins/http2_adapter/lib/src/connection_manager_imp.dart
@@ -5,6 +5,7 @@ class _ConnectionManager implements ConnectionManager {
   _ConnectionManager({
     Duration? idleTimeout,
     this.onClientCreate,
+    this.proxyConnectedPredicate = defaultProxyConnectedPredicate,
   }) : _idleTimeout = idleTimeout ?? const Duration(seconds: 1);
 
   /// Callback when socket created.
@@ -12,6 +13,9 @@ class _ConnectionManager implements ConnectionManager {
   /// We can set trusted certificates and handler
   /// for unverifiable certificates.
   final void Function(Uri uri, ClientSetting)? onClientCreate;
+
+  /// {@macro dio_http2_adapter.ProxyConnectedPredicate}
+  final ProxyConnectedPredicate proxyConnectedPredicate;
 
   /// Sets the idle timeout(milliseconds) of non-active persistent
   /// connections. For the sake of socket reuse feature with http/2,
@@ -103,6 +107,7 @@ class _ConnectionManager implements ConnectionManager {
         uri.port,
       );
       if (!isCertApproved) {
+        // TODO(EVERYONE): Replace with DioException.badCertificate once upgrade dependencies Dio above 5.4.2.
         throw DioException(
           requestOptions: options,
           type: DioExceptionType.badCertificate,
@@ -150,7 +155,7 @@ class _ConnectionManager implements ConnectionManager {
           timeout: timeout,
         );
       }
-      return SecureSocket.connect(
+      final socket = await SecureSocket.connect(
         target.host,
         target.port,
         timeout: timeout,
@@ -158,6 +163,8 @@ class _ConnectionManager implements ConnectionManager {
         onBadCertificate: clientConfig.onBadCertificate,
         supportedProtocols: ['h2'],
       );
+      _throwIfH2NotSelected(target, socket);
+      return socket;
     }
 
     final proxySocket = await Socket.connect(
@@ -173,7 +180,9 @@ class _ConnectionManager implements ConnectionManager {
     // Use CRLF as the end of the line https://www.ietf.org/rfc/rfc2616.txt
     const crlf = '\r\n';
 
-    proxySocket.write('CONNECT ${target.host}:${target.port} HTTP/1.1');
+    // TODO(EVERYONE): Figure out why we can only use an HTTP/1.x proxy here.
+    const proxyProtocol = 'HTTP/1.1';
+    proxySocket.write('CONNECT ${target.host}:${target.port} $proxyProtocol');
     proxySocket.write(crlf);
     proxySocket.write('Host: ${target.host}:${target.port}');
 
@@ -203,18 +212,25 @@ class _ConnectionManager implements ConnectionManager {
         final response = ascii.decode(event);
         final lines = response.split(crlf);
         final statusLine = lines.first;
-
-        if (statusLine.startsWith('HTTP/1.1 200')) {
-          completerProxyInitialization.complete();
-        } else {
-          completerProxyInitialization.completeError(
-            SocketException('Proxy cannot be initialized'),
-          );
+        if (!completerProxyInitialization.isCompleted) {
+          if (proxyConnectedPredicate(proxyProtocol, statusLine)) {
+            completerProxyInitialization.complete();
+          } else {
+            completerProxyInitialization.completeError(
+              SocketException(
+                'Proxy cannot be initialized with status = [$statusLine], '
+                'host = ${target.host}, port = ${target.port}',
+              ),
+            );
+          }
         }
       },
-      onError: completerProxyInitialization.completeError,
+      onError: (e, s) {
+        if (!completerProxyInitialization.isCompleted) {
+          completerProxyInitialization.completeError(e, s);
+        }
+      },
     );
-
     await completerProxyInitialization.future;
 
     final socket = await SecureSocket.secure(
@@ -224,6 +240,7 @@ class _ConnectionManager implements ConnectionManager {
       onBadCertificate: clientConfig.onBadCertificate,
       supportedProtocols: ['h2'],
     );
+    _throwIfH2NotSelected(target, socket);
 
     proxySubscription.cancel();
 
@@ -249,6 +266,12 @@ class _ConnectionManager implements ConnectionManager {
     _forceClosed = force;
     if (force) {
       _transportsMap.forEach((key, value) => value.dispose());
+    }
+  }
+
+  void _throwIfH2NotSelected(Uri target, SecureSocket socket) {
+    if (socket.selectedProtocol != 'h2') {
+      throw DioH2NotSupportedException(target, socket.selectedProtocol);
     }
   }
 }